identity/requirements.txt
asgiref==3.4.1
attrs==21.2.0
cycler==0.11.0
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.20.2
inflection==0.5.1
jsonschema==4.1.2
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
psycopg2-binary==2.9.1
PyJWT==2.3.0
pyparsing==3.0.5
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
